qcom: msm: event-timer: Add eventimer migration framework
Currently, all the event timers are associated to CPU0. These event
timers are not bind to any of the IRQs and always handled(hrtimer)
by CPU0. This event timers framework make sure event timers are
bound to clients IRQs and when the IRQ affinity has changed event
timer framework will make sure to move associated event timer to
a new CPU based on IRQ affinity.

CRs-fixed: 719258
Change-Id: Idbce3d5195655f558f74c208216f75515d235b8e
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>

qcom: msm: event-timer: Check the cpu num before adding a new event-timer

The cpu on which the new event timer is setup is calculated from
the cpumask data sent to the api. Add a check to make sure that
the calculated cpu number is within bounds. If the cpu number is
out of bounds, use the first online cpu to setup the event timer.

Also add the same bounds check to the cpu used in the irq affinity
change notifier.

Change-Id: I9f552ec506f0e69188a60884c4d820782196e5cf
Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/event_timer.c