ASoC: wcd: avoid the fake button release of headset hook
When we long press the headset hook key and then the system
resumes from suspend mode, it will report the release event
immediately from the input system level which we call a fake
release. It will mislead the userspace layer to make the
incorrect decision on the long or short press.
So we set the INPUT_PROP_NO_DUMMY_RELEASE in property bit of the
Button Jack to avoid this fake release.

Signed-off-by: Walter Yang <yandongy@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Conflicts:
	sound/soc/codecs/wcd9xxx-mbhc.c