ASoC: msm: qdsp6v2: make frame info to static array
Frame info which track size of captured data and offset is dyanmically
allocated data. Some crashes are reported due to frame info is used after
free. Error type is null pointer dereference. Total size of frame info is
relatively small and it is part of msm_audio which is dynamically
allocated structure. No need to allocate and track it separately.

Change-Id: I393a935cb43454af7b208039bf39af7bfe7feb06
Signed-off-by: Fred Oh <fred@codeaurora.org>