input: synaptics_dsx_i2c: Ensure I2C bus is active inside ISR thread
When the device is suspended, irq_wake is apparently not enough to ensure
that the I2C bus is active while an interrupt is processed.

In order to prevent gestures from failing during power collapse, check if
the I2C bus is active inside the ISR thread. If it isn't, wait until it
becomes active before processing the interrupt (typically a ~10ms wait).

This fixes gestures not working sometimes due to the following errors:
[  533.370148] qup_i2c f9924000.i2c: xfer not allowed when ctrl is suspended addr:0x20
[  533.400084] synaptics-rmi-ts 2-0020: synaptics_rmi4_i2c_read: I2C read over retry limit
[  533.400113] synaptics-rmi-ts 2-0020: synaptics_rmi4_sensor_report: Failed to read interrupt status

Also, use a dedicated wakeup source for the ISR thread to prevent the suspend
routines from interfering.

Signed-off-by: Sultanxda <sultanxda@gmail.com>